backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
python-multipart>=0.0.9
openai-whisper
sentence-transformers>=2.5.1
faiss-cpu>=1.8.0
numpy>=1.26.4
torch>=2.2.1
moviepy>=1.0.3
python-dotenv>=1.0.1
pydantic-settings>=2.2.1
ffmpeg-python>=0.2.0
setuptools
